+++ test/test_components.py
@@ -37,7 +37,7 @@ from gradio.components.dataframe import DataframeData
 from gradio.components.file_explorer import FileExplorerData
 from gradio.components.image_editor import EditorData
 from gradio.components.video import VideoData
-from gradio.data_classes import FileData, ListFiles
+from gradio.data_classes import FileData, GradioModel, GradioRootModel, ListFiles
 
 os.environ["GRADIO_ANALYTICS_ENABLED"] = "False"
 
@@ -1543,6 +1543,7 @@ class TestVideo:
                 "size": None,
                 "url": None,
                 "is_stream": False,
+                "meta": {"_type": "gradio.FileData"},
             },
             "subtitles": None,
         }
@@ -1555,6 +1556,7 @@ class TestVideo:
                 "size": None,
                 "url": None,
                 "is_stream": False,
+                "meta": {"_type": "gradio.FileData"},
             },
             "subtitles": {
                 "path": "s1.srt",
@@ -1563,6 +1565,7 @@ class TestVideo:
                 "size": None,
                 "url": None,
                 "is_stream": False,
+                "meta": {"_type": "gradio.FileData"},
             },
         }
         postprocessed_video["video"]["path"] = os.path.basename(
@@ -2262,6 +2265,7 @@ class TestGallery:
                     "size": None,
                     "url": url,
                     "is_stream": False,
+                    "meta": {"_type": "gradio.FileData"},
                 },
                 "caption": None,
             }
@@ -2290,6 +2294,7 @@ class TestGallery:
                     "size": None,
                     "url": None,
                     "is_stream": False,
+                    "meta": {"_type": "gradio.FileData"},
                 },
                 "caption": "foo_caption",
             },
@@ -2301,6 +2306,7 @@ class TestGallery:
                     "size": None,
                     "url": None,
                     "is_stream": False,
+                    "meta": {"_type": "gradio.FileData"},
                 },
                 "caption": "bar_caption",
             },
@@ -2312,6 +2318,7 @@ class TestGallery:
                     "size": None,
                     "url": None,
                     "is_stream": False,
+                    "meta": {"_type": "gradio.FileData"},
                 },
                 "caption": None,
             },
@@ -2323,6 +2330,7 @@ class TestGallery:
                     "size": None,
                     "url": None,
                     "is_stream": False,
+                    "meta": {"_type": "gradio.FileData"},
                 },
                 "caption": None,
             },
@@ -2462,7 +2470,7 @@ class TestScatterPlot:
             "elem_classes": [],
             "interactive": None,
             "label": None,
-            "name": "plot",
+            "name": "scatterplot",
             "bokeh_version": "3.0.3",
             "show_actions_button": False,
             "proxy_url": None,
@@ -2622,7 +2630,7 @@ class TestLinePlot:
             "elem_classes": [],
             "interactive": None,
             "label": None,
-            "name": "plot",
+            "name": "lineplot",
             "bokeh_version": "3.0.3",
             "show_actions_button": False,
             "proxy_url": None,
@@ -2727,7 +2735,7 @@ class TestBarPlot:
             "elem_classes": [],
             "interactive": None,
             "label": None,
-            "name": "plot",
+            "name": "barplot",
             "bokeh_version": "3.0.3",
             "show_actions_button": False,
             "proxy_url": None,
@@ -2977,3 +2985,25 @@ def test_component_example_values(io_components):
         else:
             c: Component = component()
         c.postprocess(c.example_value())
+
+
+def test_component_example_payloads(io_components):
+    for component in io_components:
+        if component == PDF:
+            continue
+        elif component in [gr.BarPlot, gr.LinePlot, gr.ScatterPlot]:
+            c: Component = component(x="x", y="y")
+        else:
+            c: Component = component()
+        data = c.example_payload()
+        data = processing_utils.move_files_to_cache(
+            data,
+            c,
+            check_in_upload_folder=False,
+        )
+        if getattr(c, "data_model", None) and data is not None:
+            if issubclass(c.data_model, GradioModel):  # type: ignore
+                data = c.data_model(**data)  # type: ignore
+            elif issubclass(c.data_model, GradioRootModel):  # type: ignore
+                data = c.data_model(root=data)  # type: ignore
+        c.preprocess(data)