+++ js/gallery/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @gradio/gallery
 
+## 0.8.2
+
+### Dependency updates
+
+- @gradio/upload@0.7.7
+- @gradio/client@0.13.0
+- @gradio/file@0.5.6
+- @gradio/image@0.9.6
+
 ## 0.8.1
 
 ### Patch Changes